+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Changed
+* Pyarrow >=15 required
+* Strawberry >=0.212 compatible
 
 ## [1.4](https://pypi.org/project/graphique/1.4/) - 2023-11-05
 ### Changed