+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 ### Changed
 * Pyarrow >=15 required
+
+### Fixed
 * Strawberry >=0.212 compatible
+* Starlette >=0.36 compatible
 
 ## [1.4](https://pypi.org/project/graphique/1.4/) - 2023-11-05
 ### Changed