+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Changed
+* `group` optimized for datasets
 
 ## [1.5](https://pypi.org/project/graphique/1.5/) - 2024-01-24
 ### Changed