+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+
+## [1.6](https://pypi.org/project/graphique/1.6/) - 2024-04-30
 ### Changed
 * Pyarrow >=16 required
 * `group` optimized for datasets